Fix test by removing sqlite3 and only testing passthrough
